moved everything to new utility class
